+++ client/package.json
@@ -13,6 +13,7 @@
     "postcss": "^8.4.33",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.0",
     "react-scripts": "5.0.1",
     "tailwindcss": "^3.4.1",
     "web-vitals": "^2.1.4"
